Refactor - Permission decorator tests (#48)
