+++ server/app/utils.py
@@ -2,9 +2,13 @@ import logging
 import smtplib
 import ssl
 from email.mime.text import MIMEText
+from functools import wraps
+from typing import Callable, List
 
 import requests
-from app.models import Settings
+from app import db
+from app.models import XSS, Client, Settings, User
+from flask_jwt_extended import get_current_user
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
@@ -90,3 +94,28 @@ def send_webhook(xss=None, receiver=None):
         requests.post(
             url=receiver, json={"text": "This is a test webhook from XSS catcher. If you are getting this, it's because your webhook configuration works."}
         )
+
+
+def permissions(all_of: List[str] = [], one_of: List[str] = []):
+    def decorator(original_function: Callable):
+        @wraps(original_function)
+        def new_function(*args, **kwargs):
+            current_user: User = get_current_user()
+
+            permission_attributes = [current_user.is_admin]
+
+            if "client_id" in kwargs:
+                client: Client = db.session.query(Client).filter_by(id=kwargs["client_id"]).first_or_404()
+                permission_attributes.append(current_user.id == client.owner_id)
+            elif "xss_id" in kwargs:
+                xss: XSS = db.session.query(XSS).filter_by(id=kwargs["xss_id"]).first_or_404()
+                permission_attributes.append(current_user.id == xss.client.owner_id)
+
+            if (all_of and all(permission_attributes)) or (one_of and any(permission_attributes)) or (not all_of and not one_of):
+                return original_function(*args, **kwargs)
+
+            return {"msg": "Forbidden"}, 403
+
+        return new_function
+
+    return decorator
