+++ server/tests/app/api/test_auth.py
