+++ server/tests/app/api/test_client.py
