+++ server/tests/app/api/test_settings.py
